Download images with skopeo instead of podman
This should make it possible to run the tool inside non-privileged
containers at the cost of slightly higher complexity.

The image is downloaded into a temporary directory, and then files in
known rpmdb locations extracted from each layer sequantially.

Since skopeo always downloads the image, the `--pull` option is no
longer useful. It is still accepted, but doesn't do anything.

This also unblocks a possible caching to avoid downloading the same
image multiple times. We could find out the digest of the image with
`skopeo inspect` first, use it as a cache key, and only download and
extract if that cache entry doesn't exist. However, this is not
implemented in this patch.