Stop depending on DNF
It is still needed, we just let the user install it manually. That was
actually always needed, since the dnf package on DNF is just a stub that
prints instructions on what to do.

Except in certain conditions, it causes pip to break down and start
spewing a lot of warnings about aborting installation, which is
massively confusing.