Add a way to list packages to reinstall
Base containers typically do not contain documentation. If a container
image wants that, it needs to reinstall the package in question.

This tool however did not support that. If a required package was
already installed, nothing would happen and it would not be listed in
the final lockfile.

This patch adds a new key to the input file. The expected value is a
list of strings with package specs, which will be reinstalled.

This can fail in two new ways
 * reinstalling something that is not in the base image
 * reinstalling when the repos no longer contain the same version

The feature only really makes sense for the container image use case,
but it can be used for the local and bare modes too.

Fixes: #12