Ignore image tag when digest is provided too
When image specification contains both a tag and a digest, remove the tag and
ignore it. Skopeo fails when both are provided, and this new behaviour
matches what podman 4.9.4 does. There's a message provided that the tag is
being stripped out.

A test case for this is added. It can be invoked by running `pytest`.

Fixes: #15