Remove URL from lockfiles
Fedora derivatives use the concept of a "compose" which is
just a bunch of JSON wrapping RPMs stored in a datestamped directory.

Including the URL of the package in the lockfile causes
total churn of the entire file anytime we bump a compose.

For the systems targeted by this tool, NEVRA uniqueness is
enforced by koji. We do not need to encode the URL.

Signed-off-by: Colin Walters <walters@verbum.org>