Fix extracting RPMDB from container images
The original code did extract the database, but did not ensure that it
is consumable by local DNF.

If the _dbpath configuration differed between local system and the
container image, the dependency resolution would run as if the database
was empty with nothing installed.

This patch fixes that by creating a symlink in such case.