Add options to select stage in containerfile
The code used to default to the first found stage, then it switched to
the last one. But the real requirement is "stage that does RPM
installation". No default can match that.

This commit adds the ability to specify the containerfile not just by
path, but also by an object with path and additional stage selectors.