Enable explicitly enabling and disabling modules
This should now cover most of modularity needs. Modules can now be
enabled, and then packages explicitly installed; or they can be
installed outright.

The tool will not inherit enabled modules from the base image, but it
still makes it possible to disable default module streams. This may not
be particularly useful, but is added for completeness.