Add caching for downloaded rpmdb
If the same image is used multiple times, we don't need to download the
rpmdb every time. This patch adds caching of the extracted rpmdb into
`~/.cache/rpm-lockfile-prototype/rpmdbs` based on the image digest.

There is no cleanup of the cache. It will grow forever and users need
to manually clean it up.

The cache is keyed by architecture and the sha256 digest of the image.
This means that even if a different pull spec is used, the cache may be
used. Unless there are collisions in the digests, this should not be a
big deal.

Fixes: #70