Add a caching wrapper
This new executable has identical interface as the original code. It
calculates a digest of all inputs (arguments and input file contents),
and will only run the actual resolver once for given inputs.

This can be helpful if you need to run the tool multiple times and want
to guarantee identical results.

The cache is never cleaned up.