Allow configuring weak dependencies
This patch adds a new top level key to the input file that allows user
to explicitly enable or disable weak dependencies rather than just
relying on the system default.

Alternatively, we could expose complete DNF config, which is not much
more complicated. However, there aren't that many other options that
could be useful for this tool.