Add experimental support for partial updates
When the `--update foo` option is passed on the command line, read
existing lockfile and only update RPMs produced by `foo` SRPM.

When running in this mode, `packages` and `reinstallPackages` sections
in the input file are ignored. The update is done by running a
transaction with all packages from the existing lockfile. The ones that
should be updated are specified by name and lower version bound only,
others are specified by full NVR and thus locked to that exact version.

If there are any dependencies with version constraints from the updated
packages to the locked packages, there will be an error. If there are no
updates available, there will also be an error.

In order for this to work, the provided repositories must contain all of
the locked packages. However, nothing requires that the same repos must
be used for the original lockfile and the update.