Add support for updating packages
A list of packages can be provided, and the equivalent of `dnf update`
will be done for each of them.

At the same time, packages to reinstall that are not available are
ignored if they are also meant to be updated. This way, you can list the
same package in `updatePackages` and `reinstallPackages`. If a newer
version is available, it will be installed. Otherwise the same version
will be reinstalled.