Ensure arches in lockfile are sorted
The arch blocks in the lockfile should use stable order. Otherwise it
would be confusing to users who could see diffs in functionally
identical files.

Processing the futures in the order of creation should not have major
performance penalty, as we need to wait on all of them anyway and
there's minimal processing when the future is done.