Use persistent cache for DNF if requested
This allows it to reuse repodata instead of always downloading it fresh.
When running on a different input with the same name repository using a
different URL, DNF notices this and re-downloads the repodata.

The user must explicitly request this by setting an env var with the
cache path. The user is then responsible for eventually cleaning this
up. The reason for this opt-in is that this feature prevents multiple
instances of the tool working in parallel.