Retry resolving with filelists if needed
DNF doesn't download filelists by default. They are generally large and
not needed. But when they are needed, there was no way of requesting
them, plus it is not easy for a user to figure that missing filelists
are the root cause.

This patch checks the exception from DNF and if it looks like missing
filelists, it will retry the transaction with filelists explicitly
enabled.

This way, most users do not have to pay the price for the extra
download. The retry reuses the same cache, so all of the other repodata
are not downloaded again.